+++ src/linktransformer/modified_sbert/losses.py
@@ -1,11 +1,14 @@
 import math
 import torch
 import torch.nn as nn
-from sentence_transformers import LoggingHandler
+from sentence_transformers import LoggingHandler, losses, SentenceTransformer
 import logging
 from transformers.utils import logging as lg
 import wandb
-
+from torch import Tensor
+from torch.nn import functional as F
+from typing import Dict,Iterable
+from enum import Enum
 
 lg.set_verbosity_error()
 logging.basicConfig(format='%(asctime)s - %(message)s',
@@ -15,7 +18,48 @@ logging.basicConfig(format='%(asctime)s - %(message)s',
 logger = logging.getLogger(__name__)
 
 
+class SiameseDistanceMetric(Enum):
+    """
+    The metric for the contrastive loss
+    """
 
+    EUCLIDEAN = lambda x, y: F.pairwise_distance(x, y, p=2)
+    MANHATTAN = lambda x, y: F.pairwise_distance(x, y, p=1)
+    COSINE_DISTANCE = lambda x, y: 1 - F.cosine_similarity(x, y)
+
+
+class OnlineContrastiveLoss_wandb(nn.Module):
+    
+    def __init__(
+        self, model: SentenceTransformer, distance_metric=SiameseDistanceMetric.COSINE_DISTANCE, margin: float = 0.5,
+        wandb_names=None):
+        super(OnlineContrastiveLoss_wandb, self).__init__()
+        self.model = model
+        self.margin = margin
+        self.distance_metric = distance_metric
+        self.wandb_names = wandb_names
+
+        
+    def forward(self, sentence_features: Iterable[Dict[str, Tensor]], labels: Tensor, size_average=False):
+        embeddings = [self.model(sentence_feature)['sentence_embedding'] for sentence_feature in sentence_features]
+
+        distance_matrix = self.distance_metric(embeddings[0], embeddings[1])
+        negs = distance_matrix[labels == 0]
+        poss = distance_matrix[labels == 1]
+
+        # select hard positive and hard negative pairs
+        negative_pairs = negs[negs < (poss.max() if len(poss) > 1 else negs.mean())]
+        positive_pairs = poss[poss > (negs.min() if len(negs) > 1 else poss.mean())]
+
+        positive_loss = positive_pairs.pow(2).sum()
+        negative_loss = F.relu(self.margin - negative_pairs).pow(2).sum()
+        loss = positive_loss + negative_loss
+
+        if self.wandb_names is not None:
+            wandb.log({"Loss": loss}) 
+
+        return loss
+